iio: imu: adis16400: move adis16400.h into adis16400.c
Final part of the re-organization. Move the `adis16400.h` into
`adis16400.c` since it no longer needs to be shared among other files.
